third_party/WebKit/LayoutTests/http/tests/workers/worker-importScripts-onerror-crossorigin-expected.txt

Ensure that scripts imported into a Worker from cross-origin hosts trigger sanitized onerror messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS onerrorMessage is "Uncaught undefined"
PASS onerrorURL is "http://127.0.0.1:8000/workers/resources/worker-importscripts-onerror-crossorigin.js"
PASS onerrorLine is 1
PASS successfullyParsed is true

TEST COMPLETE
